Started separating gui into OOP parts to begin refactoring of main menu class.
