+++ gui/views/frames/__init__.py
@@ -1,0 +1,2 @@
+from .header import CustomHeader
+from .body import CustomBody
