+++ gui/views/frames/body.py
@@ -1,0 +1,8 @@
+from .frame import CustomFrame
+
+
+class CustomBody(CustomFrame):
+    def __init__(self, frame, **kwargs):
+        super().__init__(
+            frame, **kwargs
+        )
