+++ gui/widgets/__init__.py
@@ -1,0 +1,1 @@
+from .button import CustomButton
